Migrate openhim interface to its own module
