+++ lib/app.js
@@ -6,7 +6,7 @@ var fs = require("fs");
 const _ = require('lodash');
 
 var config = JSON.parse(fs.readFileSync("config/default.json"));
-var channel = require('./channel.js');
+var OpenHIM = require('./openhim.js');
 
 exports.getDocuments = getDocuments = function(callback) {
   http.get(config.openinfoman, (res) => {
@@ -44,21 +44,25 @@ exports.createChannelConfigObject = createChannelConfigObject = function(name, c
   callback(null, channelConfig);
 }
 
-exports.registerChannel = channel.registerChannel;
-
 module.exports = exports;
 
-exports.run = function(callback) {
+exports.run = function(options, callback) {
   if (!callback) {
     callback = function() {};
   }
+
+  const openhim = OpenHIM({
+    username: options.username,
+    password: options.password,
+    url: options.openhim
+  })
 
   console.log('Fetching OpenInfoMan Documents...');
   exports.getDocuments((err, docs) => {
     docs.forEach((doc) => {
       console.log(`Creating channel for ${doc}`);
       createChannelConfigObject(doc, (err, channel) => {
-        exports.registerChannel(channel, (err, res) => {})
+        openhim.registerChannel(channel, (err, res) => {})
       })
     })
   })
@@ -68,8 +72,8 @@ if (!module.parent) {
 
   const definitions = [
     { name: 'help', alias: 'h', type: Boolean, description: 'Display this usage guide.' },
-    { name: 'openinfoman', alias: 'i', defaultValue: 'http://localhost:8984/', description: 'The openinfoman base url' },
-    { name: 'openhim', alias: 'm', defaultValue: 'http://localhost:8080/', description: 'The OpenHIMs base API url' },
+    { name: 'openinfoman', alias: 'i', defaultValue: 'http://localhost:8984', description: 'The openinfoman base url' },
+    { name: 'openhim', alias: 'm', defaultValue: 'https://localhost:8080', description: 'The OpenHIMs base API url' },
     { name: 'username', alias: 'u', defaultValue: 'root@openhim.org', description: 'The OpenHIMs API username' },
     { name: 'password', alias: 'p', description: 'The OpenHIMs API password' }
   ];
@@ -80,7 +84,7 @@ if (!module.parent) {
     console.log(cli.getUsage({
       title: 'cnf-openhim-openinfoman',
       description: 'This script creates custom channels in the OpenHIM for each registered document in an OpenInfoMan instance',
-      synopsis: '$ cnf-openhim-openinfoman -i http://localhost:8984/ -m http://localhost:8080/ -u root@openhim.org -p password'
+      synopsis: '$ cnf-openhim-openinfoman -i http://localhost:8984 -m https://localhost:8080 -u root@openhim.org -p password'
     }));
     process.exit();
   }
